use ssh-config for a little more flexibility
This still isn't as flexible as we probably want it to be. 
It still won't be easy to maintain ssh key mappings for ephemeral servers.
But at least this new approach removes the direct dependence on a single 
identify file.

Also note there is a bug in littlechef when using a ssh-config like this
the 'fix node:$host ssh:"do something"' command doesn't work. Here
is the bug report: https://github.com/tobami/littlechef/issues/94