** this doesn’t work **
This points to a new ubuntu image. Which required some slightly different
configuration. I updated Fog to give better error messages.
With an older version of little chef the command to install Chef didn’t work.
Because it couldn’t find a package for this combination of Chef 0.10 and
Ubuntu 14
With the newest version of little chef, Chef 11 is installed which isn’t
compatible with our current cookbooks.